// src/model/SilvarProfite.ts
export interface SilvarProfite {
    SIL_PROFIT_ID?: number;
    TRAND_ID?: number;
    TRANS_DATE?: Date;
    ZAKAH_DATE?: Date;
    SPOT_PRICE_OZ?: number;
    NO_OZ?: number;
    TOT_STOCK_PRICE?: number;
    MARKET_PRICE?: number;
    SALES_TAX?: number;
    TOT_MARKET_SALE_TAX?: number;
    TRANS_COST?: number;
    VALUE_BASED_CURR_PRICE?: number;
    PROFIT_VALUE?: number;
    PROFIT_PERCENT?: number;
    ZAKAH_VALUE?: number;
}